web/requirements.txt
Flask==1.0.3
Flask-QRcode==3.0.0
psycopg2-binary==2.8.3
Flask-Login==0.4.1
flask-wtf==0.14.2
pytz==2019.1